feat: Add simple perp for assignment

+++ src/oracle/interfaces/IPriceOracle.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 interface IPriceOracle {
     /// @dev Store the oracle object to storage.
-    function setPrice(address _token, uint _price) external;
+    function setPrice(address _token, uint256 _price) external;
 
     /// @dev Get the price from storage.
     function getPrice(address _token) external view returns (uint256 price);
